pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audron-repro"
version = "0.1.0"
description = "PyTorch reproduction of AUDRON from arXiv:2512.20407v2"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "librosa>=0.10",
  "soundfile>=0.12",
  "PyYAML>=6.0",
  "scikit-learn>=1.3",
  "matplotlib>=3.8",
  "pandas>=2.0",
  "tqdm>=4.66",
  "torch>=2.1"
]

[tool.setuptools.packages.find]
where = ["src"]
